+++ src/components/Execution/index.tsx
@@ -11,7 +11,7 @@ import {
     setExecutionSidePanelCheckStatus
     // updateExecutionSidePanelCheckStatus
 } from "@app/reducer/executionSlice";
-import { useAppDispatch, useAppSelector, useExecutionTemplate } from "@app/store/hooks";
+import { useAppDispatch, useAppSelector, useExecutionTemplate, useExecutionPolling } from "@app/store/hooks";
 import withLoading from "@app/components/hocs/withLoading";
 import withErrorHandling from "@app/components/hocs/withErrorHandling";
 import ArrowBackIosRoundedIcon from "@mui/icons-material/ArrowBackIosRounded";
@@ -35,17 +35,17 @@ const WorkflowWithLoading = withLoading(Workflow);
 const WorkflowWithLoadingAndErrorHandling = withErrorHandling(WorkflowWithLoading);
 
 const ExecutionComponent: React.FC<{
-    currentExecution: DatawolfExecutionFile | null;
     create: boolean;
-    wfId: string | undefined;
-}> = ({ currentExecution, create, wfId }): JSX.Element => {
+}> = ({ create }): JSX.Element => {
     const navigate = useNavigate();
+    const { wfID } = useParams<{ exId: string; wfID: string }>();
     const appDispatch = useAppDispatch();
     const { id } = useParams<{ id: string }>();
     const { setNodes, setEdges } = useStore(useShallow(selector));
 
     const sidePanelData = useAppSelector((state) => state.execution.sidePanelData);
     // const currentWorkflow = useAppSelector((state) => state.workflow.currentWorkflow);
+    const currentExecution = useAppSelector((state) => state.execution.currentExecution);
     const reactFlowWorkflow = useAppSelector((state) => state.workflow.reactFlowWorkflow);
     const workflowLoading = useAppSelector((state) => state.workflow.workflowLoading);
     const workflowError = useAppSelector((state) => state.workflow.workflowError);
@@ -54,7 +54,10 @@ const ExecutionComponent: React.FC<{
     );
     const [openExecutionDialog, setOpenExecutionDialog] = React.useState(false);
 
-    useExecutionTemplate(wfId);
+    useExecutionTemplate(wfID);
+    if (currentExecution !== null) {
+        useExecutionPolling(currentExecution.id, 10000);
+    }
 
     const handleBackClick = (wfId: string | undefined) => {
         appDispatch(resetExecutionState());
@@ -72,8 +75,8 @@ const ExecutionComponent: React.FC<{
     React.useEffect(() => {
         if (!create && currentExecution !== null) {
             appDispatch(getWorkflow({ workflowID: currentExecution.workflowId, isExecution: true }));
-        } else if (create && wfId !== undefined) {
-            appDispatch(getWorkflow({ workflowID: wfId, isExecution: true }));
+        } else if (create && wfID !== undefined) {
+            appDispatch(getWorkflow({ workflowID: wfID, isExecution: true }));
         }
     }, []);
 
@@ -85,7 +88,7 @@ const ExecutionComponent: React.FC<{
                         <Stack direction="row" spacing={2}>
                             <Box sx={{ alignContent: "center" }}>
                                 <Tooltip title="Go back" variant="plain" color="neutral" sx={{ color: "#172B4D" }}>
-                                    <IconButton variant="plain" onClick={() => handleBackClick(wfId)}>
+                                    <IconButton variant="plain" onClick={() => handleBackClick(wfID)}>
                                         <ArrowBackIosRoundedIcon />
                                     </IconButton>
                                 </Tooltip>
@@ -205,7 +208,7 @@ const ExecutionComponent: React.FC<{
                                     if (create) {
                                         setOpenExecutionDialog(true);
                                     } else {
-                                        navigate(`/project/${id}/workflows/${wfId}/execution/create`);
+                                        navigate(`/project/${id}/workflows/${wfID}/execution/create`);
                                     }
                                 }}
                             >
@@ -231,7 +234,7 @@ const ExecutionComponent: React.FC<{
             <CreateExecutionDialog
                 open={openExecutionDialog}
                 onClose={() => setOpenExecutionDialog(false)}
-                wfId={wfId}
+                wfId={wfID}
                 id={id}
             />
 
@@ -263,7 +266,7 @@ const ExecutionComponentWithErrorHandling = withErrorHandling(ExecutionComponent
 const ExecutionComponentWithLoadingAndErrorHandling = withLoading(ExecutionComponentWithErrorHandling);
 
 const Execution: React.FC<{ create: boolean }> = ({ create }): JSX.Element => {
-    const { exId, wfID } = useParams<{ exId: string; wfID: string }>();
+    const { exId } = useParams<{ exId: string; wfID: string }>();
     const appDispatch = useAppDispatch();
     const currentExecution = useAppSelector((state) => state.execution.currentExecution);
     const loading = useAppSelector((state) => state.execution.loading);
@@ -271,19 +274,12 @@ const Execution: React.FC<{ create: boolean }> = ({ create }): JSX.Element => {
 
     React.useEffect(() => {
         if (!create && exId !== undefined && currentExecution === null) {
+            // Dispatch immediately on mount
             appDispatch(getExecutionById(exId));
         }
     }, []);
 
-    return (
-        <ExecutionComponentWithLoadingAndErrorHandling
-            wfId={wfID}
-            create={create}
-            currentExecution={currentExecution}
-            isLoading={loading}
-            error={error}
-        />
-    );
+    return <ExecutionComponentWithLoadingAndErrorHandling create={create} isLoading={loading} error={error} />;
 };
 
 export default Execution;