+++ src/store/hooks.ts
@@ -5,7 +5,7 @@ import config from "@app/app.config";
 
 import { useDispatch, useSelector } from "react-redux";
 import type { RootState, AppDispatch } from "@app/store";
-import { setCreateExecutionTemplate } from "@app/reducer/executionSlice";
+import { setCreateExecutionTemplate, setCurrentExecution } from "@app/reducer/executionSlice";
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
@@ -46,4 +46,32 @@ export const useExecutionTemplate = (
     }, [wfId]);
 
     return [createExecution, state.loading, state.error];
+};
+
+export const useExecutionPolling = (executionId: string | null, interval: number) => {
+    const appDispatch = useAppDispatch();
+
+    React.useEffect(() => {
+        if (!executionId || !interval) return;
+
+        const fetchExecution = async () => {
+            try {
+                const response = await axios.get<DatawolfExecutionFile>(
+                    `${config.datawolfApi}/executions/${executionId}`,
+                    { headers: getHeaders() }
+                );
+
+                appDispatch(setCurrentExecution(response.data));
+            } catch (error) {
+                console.error("Error fetching execution: ", error);
+            }
+        };
+
+        fetchExecution(); // initial fetch
+
+        const intervalId = setInterval(fetchExecution, interval);
+
+        // cleanup
+        return () => clearInterval(intervalId);
+    }, [executionId, interval, appDispatch]);
 };